+++ src/Repository/Model/User.php
@@ -686,7 +686,7 @@ class User extends database_object
     /**
      * update_fullname_public
      * updates their fullname public
-     * @param $new_fullname_public
+     * @param bool|string $new_fullname_public
      */
     public function update_fullname_public($new_fullname_public)
     {